% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthomap.R
\name{entrez_annotate}
\alias{entrez_annotate}
\title{Return annotation information for given Entrez IDs}
\usage{
entrez_annotate(genes, taxon = NULL, column = "SYMBOL", orgdb = NULL)
}
\arguments{
\item{genes}{a numeric vector of Entrez gene IDs}

\item{taxon}{a single value for a taxon. If NULL, orthomapper will
attempt to identify the taxon (it is much faster if you specify it)}

\item{column}{a character vector with the columns requested from the annotation database}

\item{orgdb}{name of the annotation package; inferred automatically for
species present in the output of the `speciesDBITable()` function.}
}
\value{
a data frame with `length(column)+1` columns; first column is the entrez ID
}
\description{
Return annotation information for given Entrez IDs
}
\examples{
\dontrun{
## get the SYMBOL for gene 52024, infer taxon automatically
entrez_annotate(52024)

## get more information on the gene ID 52024
entrez_annotate(52024, taxon=10090, c("SYMBOL", "GENENAME"))
}
}
