% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{get_taxon}
\alias{get_taxon}
\title{Match taxon based on a string}
\usage{
get_taxon(name)
}
\arguments{
\item{name}{character vector with species names or monikers}
}
\value{
named numeric ID vector of the taxons
}
\description{
Return an NCBI taxon ID based on species name 
or moniker (such as "Hs" for Homo sapiens or "Mm" for mouse).
Note that if multiple species match the strings, only the first one will
be returned.
}
\seealso{
species_search
}
