% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{orthologyDBFromFile}
\alias{orthologyDBFromFile}
\title{Create orthology database from flat file}
\usage{
orthologyDBFromFile(
  input_file,
  output_file,
  fields = c(1, 2, 4, 5),
  info = NULL
)
}
\arguments{
\item{input_file}{Path to input file (flat text or gzipped)}

\item{output_file}{Path to SQLite output file (must not exist!)}

\item{fields}{Columns in input_file corresponding to source taxon ID, source gene ID, target taxon ID and target gene ID}

\item{info}{named list holding additional information (comment) which should be stored in the database}
}
\description{
Convert orthology database stored as a tab separated values file to SQLite
database
}
\details{
This function was used to derive the sqlite database of orthologs
included in the orthomapper package. The original source for the orthologs
was the file gene_orthologs.gz downloaded from
the NCBI gene FTP site, https://ftp.ncbi.nlm.nih.gov/gene/DATA/.
}
\examples{
\dontrun{
download.file("https://ftp.ncbi.nlm.nih.gov/gene/DATA/gene_orthologs.gz")
orthologyDBFromFile("gene_orthologs.gz", "orthologsDB.sqlite", info=list(date=Sys.Date()))
}
}
