% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthomap.R
\name{similar_symbol}
\alias{similar_symbol}
\title{Search for symbols similar to a given string}
\usage{
similar_symbol(x, taxon, orgdb = NULL)
}
\arguments{
\item{taxon}{a single value for a taxon}

\item{orgdb}{name of the annotation package; inferred automatically for
species present in the output of the `speciesDBITable()` function.}
}
\value{
a data frame with 2 columns; first column is the entrez ID, the second the symbol, third the gene name.
}
\description{
Search for symbols similar to a given string
}
\details{
Basically, uppercase the requested symbol, remove all digits at the end
and see whether anything matches.
}
\examples{
## find a gene symbol by similarity search
similar_symbol("GBP5", 10090)
}
