% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species_search}
\alias{species_search}
\title{Find all species matching a pattern}
\usage{
species_search(
  pattern,
  columns = c("name", "common_names", "lineage"),
  ignore.case = TRUE
)
}
\arguments{
\item{pattern}{regular expression to serch for}

\item{columns}{columns from the species table to search through}

\item{ignore.case}{if TRUE (default) do a case-insensitive search}
}
\description{
Find all species matching a pattern
}
\examples{
## find Danio rerio
species_search("danio")
## find all birds
species_search("aves")
}
