% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{speciesDBITable}
\alias{speciesDBITable}
\title{Return the species DBI table}
\usage{
speciesDBITable()
}
\description{
Return the species DBI table
}
\details{
This file contains information about annotationDBI org.*db files of
which orthomapper is aware. For the taxon IDs in this database, the
respective org.*db will be loaded automatically if necessary.
}
