% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthomap.R
\name{use_orthoDB}
\alias{use_orthoDB}
\title{Connect to an alternative orthology DB}
\usage{
use_orthoDB(file)
}
\arguments{
\item{file}{path to the SQLite file with the DB}
}
\value{
invisibly returns the DBI connection to the database
}
\description{
Use an alternative orthology DB SQLite file, for example if you want a newer version.
The database must have a table "orthologs" with the defined columns
"TaxID1", "GeneID1", "TaxID2", "GeneID2".
}
\seealso{
[orthologyDBFromFile()] on how to generate orthology DB files.
}
