% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthomap.R
\name{orthologs}
\alias{orthologs}
\title{All ortholog pairs between two taxons}
\usage{
orthologs(taxon1, taxon2)
}
\arguments{
\item{taxon1}{source taxon ID}

\item{taxon2}{target taxon ID}

\item{ids}{entrez gene IDs from taxon1}
}
\value{
A data frame with two columns
}
\description{
All ortholog pairs between two taxons
}
\examples{
\dontrun{
mouserat <- orthologs(10090, 10116)
}
}
