% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthomap.R
\name{orthomap}
\alias{orthomap}
\title{Map a set of gene IDs to orthologs from another taxon}
\usage{
orthomap(genes, taxon1, taxon2)
}
\arguments{
\item{genes}{a vector with entrez IDs of genes}

\item{taxon1}{source taxon ID}

\item{taxon2}{target taxon ID}
}
\value{
A data frame with two columns
}
\description{
Map a set of gene IDs to orthologs from another taxon
}
\examples{
orthomap(c(229898, 52024), 10090, 9606)
}
