% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species_info}
\alias{species_info}
\title{Retrieve species info}
\usage{
species_info(taxonID = NULL, show.lineage = FALSE)
}
\arguments{
\item{taxonID}{numeric vector of taxonomy IDs. If NULL, retrieve all species. Otherwise, find the
species selected}

\item{show.lineage}{if TRUE, show also the lineage column (not often needed)}
}
\description{
Retrieve species info either for all species in the orthology database
or for the species requested throught the taxonID parameter.
}
\examples{
species_info(c(9606, 10090))
}
